Final code clean up: refactored code to make smaller components
